+++ sdk/dotnet/RemoteNpmRepository.cs
@@ -50,6 +50,11 @@ namespace Pulumi.Artifactory
         [Output("allowAnyHostAuth")]
         public Output<bool> AllowAnyHostAuth { get; private set; } = null!;
 
+        /// <summary>
+        /// The number of seconds the repository stays in assumed offline state after a connection error. At the end of this time,
+        /// an online check is attempted in order to reset the offline status. A value of 0 means the repository is never assumed
+        /// offline. Default to 300.
+        /// </summary>
         [Output("assumedOfflinePeriodSecs")]
         public Output<int?> AssumedOfflinePeriodSecs { get; private set; } = null!;
 
@@ -258,6 +263,11 @@ namespace Pulumi.Artifactory
         [Input("allowAnyHostAuth")]
         public Input<bool>? AllowAnyHostAuth { get; set; }
 
+        /// <summary>
+        /// The number of seconds the repository stays in assumed offline state after a connection error. At the end of this time,
+        /// an online check is attempted in order to reset the offline status. A value of 0 means the repository is never assumed
+        /// offline. Default to 300.
+        /// </summary>
         [Input("assumedOfflinePeriodSecs")]
         public Input<int>? AssumedOfflinePeriodSecs { get; set; }
 
@@ -426,6 +436,11 @@ namespace Pulumi.Artifactory
         [Input("allowAnyHostAuth")]
         public Input<bool>? AllowAnyHostAuth { get; set; }
 
+        /// <summary>
+        /// The number of seconds the repository stays in assumed offline state after a connection error. At the end of this time,
+        /// an online check is attempted in order to reset the offline status. A value of 0 means the repository is never assumed
+        /// offline. Default to 300.
+        /// </summary>
         [Input("assumedOfflinePeriodSecs")]
         public Input<int>? AssumedOfflinePeriodSecs { get; set; }