+++ sdk/go/artifactory/remoteDockerRepository.go
@@ -52,7 +52,10 @@ type RemoteDockerRepository struct {
 
 	// Also known as 'Lenient Host Authentication', Allow credentials of this repository to be used on requests redirected to
 	// any other host.
-	AllowAnyHostAuth         pulumi.BoolOutput   `pulumi:"allowAnyHostAuth"`
+	AllowAnyHostAuth pulumi.BoolOutput `pulumi:"allowAnyHostAuth"`
+	// The number of seconds the repository stays in assumed offline state after a connection error. At the end of this time,
+	// an online check is attempted in order to reset the offline status. A value of 0 means the repository is never assumed
+	// offline. Default to 300.
 	AssumedOfflinePeriodSecs pulumi.IntPtrOutput `pulumi:"assumedOfflinePeriodSecs"`
 	// (A.K.A 'Ignore Repository' on the UI) When set, the repository or its local cache do not participate in artifact
 	// resolution.
@@ -157,8 +160,11 @@ func GetRemoteDockerRepository(ctx *pulumi.Context,
 type remoteDockerRepositoryState struct {
 	// Also known as 'Lenient Host Authentication', Allow credentials of this repository to be used on requests redirected to
 	// any other host.
-	AllowAnyHostAuth         *bool `pulumi:"allowAnyHostAuth"`
-	AssumedOfflinePeriodSecs *int  `pulumi:"assumedOfflinePeriodSecs"`
+	AllowAnyHostAuth *bool `pulumi:"allowAnyHostAuth"`
+	// The number of seconds the repository stays in assumed offline state after a connection error. At the end of this time,
+	// an online check is attempted in order to reset the offline status. A value of 0 means the repository is never assumed
+	// offline. Default to 300.
+	AssumedOfflinePeriodSecs *int `pulumi:"assumedOfflinePeriodSecs"`
 	// (A.K.A 'Ignore Repository' on the UI) When set, the repository or its local cache do not participate in artifact
 	// resolution.
 	BlackedOut *bool `pulumi:"blackedOut"`
@@ -228,7 +234,10 @@ type remoteDockerRepositoryState struct {
 type RemoteDockerRepositoryState struct {
 	// Also known as 'Lenient Host Authentication', Allow credentials of this repository to be used on requests redirected to
 	// any other host.
-	AllowAnyHostAuth         pulumi.BoolPtrInput
+	AllowAnyHostAuth pulumi.BoolPtrInput
+	// The number of seconds the repository stays in assumed offline state after a connection error. At the end of this time,
+	// an online check is attempted in order to reset the offline status. A value of 0 means the repository is never assumed
+	// offline. Default to 300.
 	AssumedOfflinePeriodSecs pulumi.IntPtrInput
 	// (A.K.A 'Ignore Repository' on the UI) When set, the repository or its local cache do not participate in artifact
 	// resolution.
@@ -303,8 +312,11 @@ func (RemoteDockerRepositoryState) ElementType() reflect.Type {
 type remoteDockerRepositoryArgs struct {
 	// Also known as 'Lenient Host Authentication', Allow credentials of this repository to be used on requests redirected to
 	// any other host.
-	AllowAnyHostAuth         *bool `pulumi:"allowAnyHostAuth"`
-	AssumedOfflinePeriodSecs *int  `pulumi:"assumedOfflinePeriodSecs"`
+	AllowAnyHostAuth *bool `pulumi:"allowAnyHostAuth"`
+	// The number of seconds the repository stays in assumed offline state after a connection error. At the end of this time,
+	// an online check is attempted in order to reset the offline status. A value of 0 means the repository is never assumed
+	// offline. Default to 300.
+	AssumedOfflinePeriodSecs *int `pulumi:"assumedOfflinePeriodSecs"`
 	// (A.K.A 'Ignore Repository' on the UI) When set, the repository or its local cache do not participate in artifact
 	// resolution.
 	BlackedOut *bool `pulumi:"blackedOut"`
@@ -372,7 +384,10 @@ type remoteDockerRepositoryArgs struct {
 type RemoteDockerRepositoryArgs struct {
 	// Also known as 'Lenient Host Authentication', Allow credentials of this repository to be used on requests redirected to
 	// any other host.
-	AllowAnyHostAuth         pulumi.BoolPtrInput
+	AllowAnyHostAuth pulumi.BoolPtrInput
+	// The number of seconds the repository stays in assumed offline state after a connection error. At the end of this time,
+	// an online check is attempted in order to reset the offline status. A value of 0 means the repository is never assumed
+	// offline. Default to 300.
 	AssumedOfflinePeriodSecs pulumi.IntPtrInput
 	// (A.K.A 'Ignore Repository' on the UI) When set, the repository or its local cache do not participate in artifact
 	// resolution.