+++ sdk/go/artifactory/remoteNpmRepository.go
@@ -49,7 +49,10 @@ type RemoteNpmRepository struct {
 
 	// Also known as 'Lenient Host Authentication', Allow credentials of this repository to be used on requests redirected to
 	// any other host.
-	AllowAnyHostAuth         pulumi.BoolOutput   `pulumi:"allowAnyHostAuth"`
+	AllowAnyHostAuth pulumi.BoolOutput `pulumi:"allowAnyHostAuth"`
+	// The number of seconds the repository stays in assumed offline state after a connection error. At the end of this time,
+	// an online check is attempted in order to reset the offline status. A value of 0 means the repository is never assumed
+	// offline. Default to 300.
 	AssumedOfflinePeriodSecs pulumi.IntPtrOutput `pulumi:"assumedOfflinePeriodSecs"`
 	// (A.K.A 'Ignore Repository' on the UI) When set, the repository or its local cache do not participate in artifact
 	// resolution.
@@ -147,8 +150,11 @@ func GetRemoteNpmRepository(ctx *pulumi.Context,
 type remoteNpmRepositoryState struct {
 	// Also known as 'Lenient Host Authentication', Allow credentials of this repository to be used on requests redirected to
 	// any other host.
-	AllowAnyHostAuth         *bool `pulumi:"allowAnyHostAuth"`
-	AssumedOfflinePeriodSecs *int  `pulumi:"assumedOfflinePeriodSecs"`
+	AllowAnyHostAuth *bool `pulumi:"allowAnyHostAuth"`
+	// The number of seconds the repository stays in assumed offline state after a connection error. At the end of this time,
+	// an online check is attempted in order to reset the offline status. A value of 0 means the repository is never assumed
+	// offline. Default to 300.
+	AssumedOfflinePeriodSecs *int `pulumi:"assumedOfflinePeriodSecs"`
 	// (A.K.A 'Ignore Repository' on the UI) When set, the repository or its local cache do not participate in artifact
 	// resolution.
 	BlackedOut *bool `pulumi:"blackedOut"`
@@ -211,7 +217,10 @@ type remoteNpmRepositoryState struct {
 type RemoteNpmRepositoryState struct {
 	// Also known as 'Lenient Host Authentication', Allow credentials of this repository to be used on requests redirected to
 	// any other host.
-	AllowAnyHostAuth         pulumi.BoolPtrInput
+	AllowAnyHostAuth pulumi.BoolPtrInput
+	// The number of seconds the repository stays in assumed offline state after a connection error. At the end of this time,
+	// an online check is attempted in order to reset the offline status. A value of 0 means the repository is never assumed
+	// offline. Default to 300.
 	AssumedOfflinePeriodSecs pulumi.IntPtrInput
 	// (A.K.A 'Ignore Repository' on the UI) When set, the repository or its local cache do not participate in artifact
 	// resolution.
@@ -279,8 +288,11 @@ func (RemoteNpmRepositoryState) ElementType() reflect.Type {
 type remoteNpmRepositoryArgs struct {
 	// Also known as 'Lenient Host Authentication', Allow credentials of this repository to be used on requests redirected to
 	// any other host.
-	AllowAnyHostAuth         *bool `pulumi:"allowAnyHostAuth"`
-	AssumedOfflinePeriodSecs *int  `pulumi:"assumedOfflinePeriodSecs"`
+	AllowAnyHostAuth *bool `pulumi:"allowAnyHostAuth"`
+	// The number of seconds the repository stays in assumed offline state after a connection error. At the end of this time,
+	// an online check is attempted in order to reset the offline status. A value of 0 means the repository is never assumed
+	// offline. Default to 300.
+	AssumedOfflinePeriodSecs *int `pulumi:"assumedOfflinePeriodSecs"`
 	// (A.K.A 'Ignore Repository' on the UI) When set, the repository or its local cache do not participate in artifact
 	// resolution.
 	BlackedOut *bool `pulumi:"blackedOut"`
@@ -341,7 +353,10 @@ type remoteNpmRepositoryArgs struct {
 type RemoteNpmRepositoryArgs struct {
 	// Also known as 'Lenient Host Authentication', Allow credentials of this repository to be used on requests redirected to
 	// any other host.
-	AllowAnyHostAuth         pulumi.BoolPtrInput
+	AllowAnyHostAuth pulumi.BoolPtrInput
+	// The number of seconds the repository stays in assumed offline state after a connection error. At the end of this time,
+	// an online check is attempted in order to reset the offline status. A value of 0 means the repository is never assumed
+	// offline. Default to 300.
 	AssumedOfflinePeriodSecs pulumi.IntPtrInput
 	// (A.K.A 'Ignore Repository' on the UI) When set, the repository or its local cache do not participate in artifact
 	// resolution.