+++ sdk/nodejs/remoteDockerRepository.ts
@@ -65,6 +65,11 @@ export class RemoteDockerRepository extends pulumi.CustomResource {
      * any other host.
      */
     public readonly allowAnyHostAuth!: pulumi.Output<boolean>;
+    /**
+     * The number of seconds the repository stays in assumed offline state after a connection error. At the end of this time,
+     * an online check is attempted in order to reset the offline status. A value of 0 means the repository is never assumed
+     * offline. Default to 300.
+     */
     public readonly assumedOfflinePeriodSecs!: pulumi.Output<number | undefined>;
     /**
      * (A.K.A 'Ignore Repository' on the UI) When set, the repository or its local cache do not participate in artifact
@@ -281,6 +286,11 @@ export interface RemoteDockerRepositoryState {
      * any other host.
      */
     allowAnyHostAuth?: pulumi.Input<boolean>;
+    /**
+     * The number of seconds the repository stays in assumed offline state after a connection error. At the end of this time,
+     * an online check is attempted in order to reset the offline status. A value of 0 means the repository is never assumed
+     * offline. Default to 300.
+     */
     assumedOfflinePeriodSecs?: pulumi.Input<number>;
     /**
      * (A.K.A 'Ignore Repository' on the UI) When set, the repository or its local cache do not participate in artifact
@@ -389,6 +399,11 @@ export interface RemoteDockerRepositoryArgs {
      * any other host.
      */
     allowAnyHostAuth?: pulumi.Input<boolean>;
+    /**
+     * The number of seconds the repository stays in assumed offline state after a connection error. At the end of this time,
+     * an online check is attempted in order to reset the offline status. A value of 0 means the repository is never assumed
+     * offline. Default to 300.
+     */
     assumedOfflinePeriodSecs?: pulumi.Input<number>;
     /**
      * (A.K.A 'Ignore Repository' on the UI) When set, the repository or its local cache do not participate in artifact