+++ sdk/python/pulumi_artifactory/remote_docker_repository.py
@@ -59,6 +59,9 @@ class RemoteDockerRepositoryArgs:
         :param pulumi.Input[str] url: - the remote repo URL. You kinda don't have a remote repo without it
         :param pulumi.Input[bool] allow_any_host_auth: Also known as 'Lenient Host Authentication', Allow credentials of this repository to be used on requests redirected to
                any other host.
+        :param pulumi.Input[int] assumed_offline_period_secs: The number of seconds the repository stays in assumed offline state after a connection error. At the end of this time,
+               an online check is attempted in order to reset the offline status. A value of 0 means the repository is never assumed
+               offline. Default to 300.
         :param pulumi.Input[bool] blacked_out: (A.K.A 'Ignore Repository' on the UI) When set, the repository or its local cache do not participate in artifact
                resolution.
         :param pulumi.Input[bool] block_mismatching_mime_types: Before caching an artifact, Artifactory first sends a HEAD request to the remote resource. In some remote resources,
@@ -198,6 +201,11 @@ class RemoteDockerRepositoryArgs:
     @property
     @pulumi.getter(name="assumedOfflinePeriodSecs")
     def assumed_offline_period_secs(self) -> Optional[pulumi.Input[int]]:
+        """
+        The number of seconds the repository stays in assumed offline state after a connection error. At the end of this time,
+        an online check is attempted in order to reset the offline status. A value of 0 means the repository is never assumed
+        offline. Default to 300.
+        """
         return pulumi.get(self, "assumed_offline_period_secs")
 
     @assumed_offline_period_secs.setter
@@ -607,6 +615,9 @@ class _RemoteDockerRepositoryState:
         Input properties used for looking up and filtering RemoteDockerRepository resources.
         :param pulumi.Input[bool] allow_any_host_auth: Also known as 'Lenient Host Authentication', Allow credentials of this repository to be used on requests redirected to
                any other host.
+        :param pulumi.Input[int] assumed_offline_period_secs: The number of seconds the repository stays in assumed offline state after a connection error. At the end of this time,
+               an online check is attempted in order to reset the offline status. A value of 0 means the repository is never assumed
+               offline. Default to 300.
         :param pulumi.Input[bool] blacked_out: (A.K.A 'Ignore Repository' on the UI) When set, the repository or its local cache do not participate in artifact
                resolution.
         :param pulumi.Input[bool] block_mismatching_mime_types: Before caching an artifact, Artifactory first sends a HEAD request to the remote resource. In some remote resources,
@@ -733,6 +744,11 @@ class _RemoteDockerRepositoryState:
     @property
     @pulumi.getter(name="assumedOfflinePeriodSecs")
     def assumed_offline_period_secs(self) -> Optional[pulumi.Input[int]]:
+        """
+        The number of seconds the repository stays in assumed offline state after a connection error. At the end of this time,
+        an online check is attempted in order to reset the offline status. A value of 0 means the repository is never assumed
+        offline. Default to 300.
+        """
         return pulumi.get(self, "assumed_offline_period_secs")
 
     @assumed_offline_period_secs.setter
@@ -1211,6 +1227,9 @@ class RemoteDockerRepository(pulumi.CustomResource):
         :param pulumi.ResourceOptions opts: Options for the resource.
         :param pulumi.Input[bool] allow_any_host_auth: Also known as 'Lenient Host Authentication', Allow credentials of this repository to be used on requests redirected to
                any other host.
+        :param pulumi.Input[int] assumed_offline_period_secs: The number of seconds the repository stays in assumed offline state after a connection error. At the end of this time,
+               an online check is attempted in order to reset the offline status. A value of 0 means the repository is never assumed
+               offline. Default to 300.
         :param pulumi.Input[bool] blacked_out: (A.K.A 'Ignore Repository' on the UI) When set, the repository or its local cache do not participate in artifact
                resolution.
         :param pulumi.Input[bool] block_mismatching_mime_types: Before caching an artifact, Artifactory first sends a HEAD request to the remote resource. In some remote resources,
@@ -1437,6 +1456,9 @@ class RemoteDockerRepository(pulumi.CustomResource):
         :param pulumi.ResourceOptions opts: Options for the resource.
         :param pulumi.Input[bool] allow_any_host_auth: Also known as 'Lenient Host Authentication', Allow credentials of this repository to be used on requests redirected to
                any other host.
+        :param pulumi.Input[int] assumed_offline_period_secs: The number of seconds the repository stays in assumed offline state after a connection error. At the end of this time,
+               an online check is attempted in order to reset the offline status. A value of 0 means the repository is never assumed
+               offline. Default to 300.
         :param pulumi.Input[bool] blacked_out: (A.K.A 'Ignore Repository' on the UI) When set, the repository or its local cache do not participate in artifact
                resolution.
         :param pulumi.Input[bool] block_mismatching_mime_types: Before caching an artifact, Artifactory first sends a HEAD request to the remote resource. In some remote resources,
@@ -1521,6 +1543,11 @@ class RemoteDockerRepository(pulumi.CustomResource):
     @property
     @pulumi.getter(name="assumedOfflinePeriodSecs")
     def assumed_offline_period_secs(self) -> pulumi.Output[Optional[int]]:
+        """
+        The number of seconds the repository stays in assumed offline state after a connection error. At the end of this time,
+        an online check is attempted in order to reset the offline status. A value of 0 means the repository is never assumed
+        offline. Default to 300.
+        """
         return pulumi.get(self, "assumed_offline_period_secs")
 
     @property